+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.0.2](https://github.com/codeskills-dev/md-to-react-email/compare/v3.0.1...v3.0.2) (2023-07-04)
+
+### Features
+
+- Added checks to handle `undefined | null | ''`
+- Added checks to handle input that is not of type `string`
+
 ### [3.0.1](https://github.com/codeskills-dev/md-to-react-email/compare/v3.0.0...v3.0.1) (2023-07-04)
 
 ### Features